--- syscall.go
@@ -1,9 +1,0 @@
-// +build !linux
-
-package bolt
-
-import "os"
-
-func fdatasync(f *os.File) error {
-	return f.Sync()
-}
